# src/swagger.yml
openapi: "3.0.0"
info:
  version: "1.0.0"
  title: "Library Management API"

paths:
  /books:
    get:
      summary: "List all books"
      responses:
        '200':
          description: "A list of books"
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Book'
    post:
      summary: "Add a new book"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/BookInput'
      responses:
        '201':
          description: "Book created"

  /books/search:
    get:
      summary: "Search for books"
      parameters:
        - name: author
          in: query
          schema:
            type: string
        - name: year
          in: query
          schema:
            type: string
        - name: genre
          in: query
          schema:
            type: string
      responses:
        '200':
          description: "Filtered list of books"

  /books/{isbn}:
    delete:
      summary: "Delete a book by ISBN"
      parameters:
        - name: isbn
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: "Book deleted"
        '404':
          description: "Book not found"

    put:
      summary: "Update a book by ISBN"
      parameters:
        - name: isbn
          in: path
          required: true
          schema:
            type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/BookInputPartial'
      responses:
        '200':
          description: "Book updated"
        '404':
          description: "Book not found"

components:
  schemas:
    Book:
      type: object
      properties:
        title:
          type: string
        author:
          type: string
        publishedYear:
          type: integer
        isbn:
          type: string
        genre:
          type: string
    BookInput:
      type: object
      required: [title, author, publishedYear, isbn]
      properties:
        title:
          type: string
        author:
          type: string
        publishedYear:
          type: integer
        isbn:
          type: string
        genre:
          type: string
    BookInputPartial:
      type: object
      properties:
        title:
          type: string
        author:
          type: string
        publishedYear:
          type: integer
        genre:
          type: string